refactor(model): refactor model buidler
