+++ internlm/model/modeling_moe.py
@@ -20,10 +20,8 @@ from internlm.model.modules.norm import new_layer_norm
 from internlm.model.modules.utils import split_forward_gather_backward
 from internlm.model.moe import MoE
 from internlm.solver.activation_checkpoint import activation_checkpoint
-from internlm.solver.pipeline_utils import partition_uniform
-from internlm.utils.common import filter_kwargs
 from internlm.utils.logger import get_logger
-from internlm.utils.registry import MODEL_INITIALIZER
+from internlm.core.model import MODEL_INITIALIZER
 
 MODEL_TYPE = "INTERNLM_MoE"
 
@@ -421,48 +419,9 @@ class PackedFlashInternLm1D(nn.Module):
         return hidden_states, moe_losses
 
 
-def _build_generic_model_1d(num_layers, num_chunks, device=torch.device("cuda"), **kwargs):
-    """
-    build generic model 1d
-
-    Args:
-        num_layers (int): The number of layer.
-        num_chunks (int): The number of partitions in pipeline parallel.
-        device (Optional[Union[str, torch.device]]): The device will be used. torch.device("cuda") by default.
-
-    """
-    pipeline_size = gpc.get_world_size(ParallelMode.PIPELINE)
-    pipeline_rank = gpc.get_local_rank(ParallelMode.PIPELINE)
-
-    all_parts = partition_uniform(num_layers, pipeline_size, num_chunks)
-    parts = all_parts[pipeline_rank]
-    if gpc.is_rank_for_log():
-        logger.info(f"The layer sharding is {all_parts}.")
-
-    models = []
-
-    for start, end in parts:
-        kwargs["num_layers"] = end - start
-        kwargs["first"] = start == 0
-        # If there is no content in the final layer, assign the last layer.
-        kwargs["last"] = end == num_layers and len(all_parts[-1]) != 0
-        kwargs["device"] = device
-        kwargs["start_layer_idx"] = start
-        chunk = PackedFlashInternLm1D(**filter_kwargs(PackedFlashInternLm1D.__init__, kwargs)).to(device)
-
-        models.append(chunk)
-    torch.distributed.barrier()
-    if len(models) == 1:
-        model = models[0]
-    else:
-        model = nn.ModuleList(models)
-
-    return model
-
-
 @MODEL_INITIALIZER.register_module(module_name=MODEL_TYPE)
 def build_model_with_moe_cfg(
-    num_chunks=1,
+    # num_chunks=1,
     checkpoint=0.0,
     dtype=torch.float,
     embed_split_hidden=False,
@@ -485,7 +444,7 @@ def build_model_with_moe_cfg(
     dropout_selective_checkpoint=True,
     use_scaled_init: bool = True,
     use_swiglu: bool = True,
-    use_flash_attn: bool = True,
+    # use_flash_attn: bool = True,
     num_experts: int = 1,
     moe_use_residual: bool = False,  # pylint: disable=W0613
     moe_type: str = None,  # pylint: disable=W0613
@@ -526,6 +485,7 @@ def build_model_with_moe_cfg(
     """
 
     cfg = dict(
+        num_layers=num_layers,
         hidden_size=hidden_size,
         num_attention_heads=num_attention_heads,
         checkpoint=checkpoint,
@@ -546,8 +506,8 @@ def build_model_with_moe_cfg(
         dropout_selective_checkpoint=dropout_selective_checkpoint,
         use_scaled_init=use_scaled_init,
         use_swiglu=use_swiglu,
-        use_flash_attn=use_flash_attn,
+        # use_flash_attn=use_flash_attn,
         num_experts=num_experts,
     )
 
-    return _build_generic_model_1d(num_layers=num_layers, num_chunks=num_chunks, **cfg)
+    return PackedFlashInternLm1D(**cfg)
