+++ tests/test_utils/common_fixture.py
@@ -91,9 +91,9 @@ def init_naive_model():
     import internlm.model.modeling_internlm  # noqa # pylint: disable=unused-import
     import internlm.model.modeling_moe  # noqa # pylint: disable=unused-import
     from internlm.core.naive_amp import NaiveAMPModel
-    from internlm.utils.registry import MODEL_INITIALIZER
+    from internlm.core.model import create_model
 
-    model = MODEL_INITIALIZER.get_module(module_name=gpc.config.model_type)(**(init_config.model))
+    model = create_model(module_name=gpc.config.model_type, **(init_config.model))
     model = NaiveAMPModel(
         model=model,
         output_to_fp32=False,
